signup

+++ fatorial.c
@@ -1,7 +1,7 @@
 
 int fatorial_func(int* fatorial)
 {
-    int aux = 1;
+    unsigned long aux = 1;
 
     for (int i = 1; i <= *fatorial; i++)
     {
@@ -10,12 +10,3 @@ int fatorial_func(int* fatorial)
 
     return aux;
 };
-
-// int main()
-// {
-//     int x = 7;
-//     fatorial_func(&x);
-
-//     printf("Fatorial de %d: %d\n", x, fatorial_func(&x));
-// }
-
